refactor code as adviced by reviewer
